Move Avatar to Primitives
