+++ ui/DesignSystem/Component/index.js
@@ -1,5 +1,4 @@
 import AdditionalActionsDropdown from "./AdditionalActionsDropdown.svelte";
-import Avatar from "./Avatar.svelte";
 import Notification from "./Notification.svelte";
 import NotificationFaucet from "./NotificationFaucet.svelte";
 import ModalLayout from "./ModalLayout.svelte";
@@ -19,7 +18,6 @@ import UserCard from "./UserCard.svelte";
 
 export {
   AdditionalActionsDropdown,
-  Avatar,
   Notification,
   NotificationFaucet,
   ModalLayout,
