+++ ui/DesignSystem/Primitive/index.js
@@ -1,3 +1,4 @@
+import Avatar from "./Avatar.svelte";
 import Button from "./Button.svelte";
 import Caption from "./Caption.svelte";
 import Code from "./Code.svelte";
@@ -8,4 +9,15 @@ import Numeric from "./Numeric.svelte";
 import Text from "./Text.svelte";
 import Title from "./Title.svelte";
 
-export { Button, Caption, Code, Flex, Icon, Input, Numeric, Text, Title };
+export {
+  Avatar,
+  Button,
+  Caption,
+  Code,
+  Flex,
+  Icon,
+  Input,
+  Numeric,
+  Text,
+  Title
+};
